+++ Quaver/App/Controller/admin/languages.php
@@ -11,7 +11,7 @@ use Quaver\Core\LangStrings;
 
 // Check privileges
 if (!$_user->logged || !$_user->isAdmin()) {
-    header("Location: /");
+    header("Location: /login");
     exit;
 } 
 
@@ -72,7 +72,7 @@ switch ($this->getCurrentURL()) {
     case('edit'):
    	 	$this->addTwigVars('typePOST', 'edit');
         $lang = new LangStrings;
-    	$item = $lang->getFromLabel($this->url_var[2]);
+    	$item = $lang->getFromLabel($this->url->uri[1]);
     	$this->addTwigVars('item', $item);
 
         // Load template with data
@@ -81,7 +81,7 @@ switch ($this->getCurrentURL()) {
     	break;
     case('del'):
         $lang = new LangStrings;
-	    $items = $lang->getFromLabel($this->url_var[2]);
+	    $items = $lang->getFromLabel($this->url->uri[1]);
 
         foreach ($items as $item) {
             $item->delete();