+++ Quaver/Core/Core.php
@@ -10,6 +10,10 @@ namespace Quaver\Core;
 use Quaver\Core\Lang;
 use Quaver\App\Model\User;
 
+/**
+ * Core
+ * @package default
+ */
 class Core
 {
     /**